+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 CHANGELOG
 =========
 
+1.6.0
+==========
+
+* Added: Support for new export API filters
+* Added: Support for assets lists in scan settings
 
 1.5.0
 ==========